Added a dedicated section for the top bar in the UI specifications
Problem: There are multiple spots in the UI specifications where the text for
the top bar is repeated. For example, in the home screen, view dataset screen,
and validation summary screen. This isn't DRY.

Solution: Added a dedicated section in the UI ui-specifications section. We can
refer to this section from any screen that uses the top bar.

Files:
* components.puml: Created this new file which will house functions to render
components. In the future it will also house the sidebar component. Moved the
$render_top_bar procedure from screens.puml to this file and renamed it to
$top_bar
* screens.puml: The $render_top_bar function has been moved to a new
components.puml and renamed $top_bar. This new file is now imported and all
instances of the old top bar function changed to use the new one.
* top-bar.puml: Created this new file to render an image for the top bar for the
UI specifications.
* ui-specifications.qmd: Updated text