Changed the dataset upload flow in the UI specifications to take into account that the user can upload the same dataset again
Problem: Once a user has validated a dataset, in order to fix any errors they
will need to update the dataset in their local file system, and re-upload it to
the tool to re-validate the new version.

Once the new version is re-uploaded, the user will need to switch to the old
version of the dataset in the tool in order to look at the previous validations.

This is cumbersome.

Solution: If the tool detects that a new file is a new version of an existing
one, the tool will give the user the option to replace the old version.

Files:

* components.puml: Added a new sidebar component which takes into
  account that the previous validations section within it will be different if
  the current dataset has multiple versions.
* sidebar-multiple-version.puml, sidebar-no-multiple-versions.puml: New files
  that shows how the sidebar will look with a dataset that has multiple
  and single version respectively.
* screens.puml: All sidebar references now uses the new $sidebar procedure to
  render it.
  Added functions $duplicate-dataset-dialog and $replace-dataset-confirm-dialog
* duplicate-dataset-dialog.puml, replace-dataset-confirm-dialog.puml: New files that
  contain the dialogs to present to the user during the dataset upload flow when
  the tool detects that the new uploaded file is a possible duplicate.
* user-flows.puml: Updated the $upload_dataset_flow procedure with the new
  replace dataset flows
* ui-specifications.qmd: Updated text