Updated the UI specifications for the choose dataset dialog and renamed it to upload dataset dialog to match the code.
Problem: Plotly does not allow you to directly use the browser's file
input which means the code could not implement the UI specifications.
Instead, a dialog is shown where the user has to click the some text to
bring up the file input dialog or drag and drop to upload the file.

Solution: Update the UI specifications to reflect the way the dialog is
implemented in the code.

Files:
* screens.puml: Renamed choose_dataset_dialog to upload_dataset_dialog
* choose-dataset-dialog.puml, upload-dataset-dialog.puml: Renamed the
  former to the latter and changed all references to choose dataset
  dialog to upload dataset dialog.
* user-flows.puml: Updated the $upload_dataset_flow procedure to change
  references to choose dataset to upload dataset
* ui-spec: Updated the UI specifications for the choose dataset dialog
  which is now the upload datast dialog