Updated the UI specifications to allow the user to ignore a sheet when configuring an excel file
Problem: When working with an excel sheet the user could get into a
pseudo soft lock state when configuring the file if a sheet cannot be mapped
onto a valid ODM table. This can happen for example if the excel file has some
metadata sheets or sheets from multiple ODM versions.

Solution: Allow the user to ignore sheets in an excel file which means that the
tool would not process it.

Files:

* screens.puml: Updated the configure dataset dialog to add a Ignore
  sheet button in the Tables dropdown
* ui-spec.qmd:
    * Updated the text for the configure excel dataset dialog to outline
      how ignoring a sheet works.
    * Updated the text for the view dataset page to outline that the
      upload report should include the list of ignored sheets in an
      excel file.