ui-spec: use same layout for all validation links
Drops the "current" label, since it's only confusing when the dataset
revision is newer than the latest validation. This could also be
resolved by checking the dataset revision from the sidebar, and only
show "current" when it's actually the current revision, but it's not
really worth it. Best to keep it simple instead. It's also much cleaner
without "current".

Keeps the nested list design for all revisions even if it's just a
single validation in the list. This keeps it more consistent and simple,
and it doesn't seem to have any drawbacks.